HellFire Visit - TycoonServer Poof + ClearasilSkater
Adds CDM API - /user/space endpoint - unimplemented
Removes old PA API files in static that is no longer valid. Will be reorganized in future PR.

Removed TycoonServer due to Clearasil Skater not using a Redirector, instead that will be merged inside HTTP/HTTPS Servers respectively 61900 port is NOT nessecary for redundancy,
