--- SpecializedServers/TycoonServer/TycoonServer.csproj
@@ -1,34 +1,0 @@
-﻿<Project Sdk="Microsoft.NET.Sdk">
-
-  <PropertyGroup>
-    <OutputType>Exe</OutputType>
-    <TargetFrameworks>net6.0;net7.0;net8.0</TargetFrameworks>
-    <ImplicitUsings>enable</ImplicitUsings>
-    <Nullable>enable</Nullable>
-	<DockerDefaultTargetOS>Linux</DockerDefaultTargetOS>
-	<DockerfileRunArguments>-p 61900:61900</DockerfileRunArguments>
-	<ApplicationIcon>Home Tycoon.ico</ApplicationIcon>
-	<ServerGarbageCollection>true</ServerGarbageCollection>
-	<SuppressTfmSupportBuildWarnings>true</SuppressTfmSupportBuildWarnings>
-  </PropertyGroup>
-
-  <ItemGroup>
-    <Content Include="Home Tycoon.ico" />
-  </ItemGroup>
-
-  <ItemGroup>
-    <PackageReference Include="Microsoft.VisualStudio.Azure.Containers.Tools.Targets" Version="1.20.1" />
-    <PackageReference Include="Watson.Lite" Version="6.1.8" />
-  </ItemGroup>
-  
-  <ItemGroup>
-    <Content Include="static\**\*.*">
-      <CopyToOutputDirectory>Always</CopyToOutputDirectory>
-    </Content>  
-  </ItemGroup>
-  
-  <ItemGroup>
-    <ProjectReference Include="..\..\BackendServices\AuxiliaryServices\WebAPIService\WebAPIService.csproj" />
-  </ItemGroup>
-
-</Project>
