--- WebServers/HTTPSecureServerLite/static/wwwroot/eventController/MikuLiveEvent/concert_complete_judge.xml
@@ -1,10 +1,0 @@
-<xml>
-<result type="int">1</result>
-<description type="text">Success</description>
-<error_no type="int">0</error_no>
-<error_message type="text">None</error_message>
-<key type="text">concert_complete_judge</key>
-<seq type="int">2</seq>
-<resource type="int">1</resource>
-<data type="text">http://scej-home.playstation.net/eventController/MikuLiveEvent/Project_diva_x_logo.png</data>
-</xml>
