--- WebServers/HTTPSecureServerLite/static/wwwroot/eventController/MikuLiveEvent/getEventTrigger.xml
@@ -1,26 +1,0 @@
-<xml>
-<result type="int">1</result>
-<description type="text">Success</description>
-<error_no type="int">0</error_no>
-<error_message type="text">None</error_message>
-<trigger_count type="int">200</trigger_count>
-<is_active type="bool">true</is_active>
-<trigger_id type="int">1</trigger_id>
-
-<!-- Event Start Trigger --> 
-<start_year type="int">2023</start_year>
-<start_month type="int">09</start_month>
-<start_day type="int">16</start_day>
-<start_hour type="int">22</start_hour>
-<start_minutes type="int">14</start_minutes>
-<start_second type="int">00</start_second>
-
-<end_year type="int">2023</end_year>
-<end_month type="int">09</end_month>
-<end_day type="int">20</end_day>
-<end_hour type="int">05</end_hour>
-<end_minutes type="int">45</end_minutes>
-<end_second type="int">0</end_second>
-
-<trigger_flag type="int">500</trigger_flag>
-</xml>
