--- WebServers/HTTPSecureServerLite/static/wwwroot/eventController/MikuLiveEvent/getUserEventCustom.xml
@@ -1,13 +1,0 @@
-<xml>
-<result type="int">1</result>
-<description type="text">Success</description>
-<error_no type="int">0</error_no>
-<error_message type="text">None</error_message>
-<field_name type="text">playrecord</field_name>
-<field_value type="int">0</field_value>
-<update_year type="int">2023</update_year>
-<update_month type="int">09</update_month>
-<update_day type="int">12</update_day>
-<update_hour type="int">21</update_hour>
-<update_second type="int">0</update_second>
-</xml>
