--- WebServers/HTTPSecureServerLite/static/wwwroot/eventController/MikuLiveJukebox/getEventTrigger.xml
@@ -1,10 +1,0 @@
-<xml>
-<result type="int">1</result>
-<description type="text">Success</description>
-<error_no type="int">0</error_no>
-<error_message type="text">None</error_message>
-<trigger_id type="int">1</trigger_id>
-
-<trigger_flag type="int">1</trigger_flag>
-
-</xml>
