--- WebServers/HTTPSecureServerLite/static/wwwroot/eventController/MikuLiveJukebox/resources/miku_jukebox1_board.xml
@@ -1,10 +1,0 @@
-<xml>
-<result type="int">1</result>
-<description type="text">Success</description>
-<error_no type="int">0</error_no>
-<error_message type="text">None</error_message>
-<key type="text">miku_jukebox1_board</key>
-<seq type="int">2</seq>
-<resource type="int">1</resource>
-<data type="text"></data>
-</xml>
