--- WebServers/HTTPSecureServerLite/static/wwwroot/eventController/infoboard/09/jul2009.xml
@@ -1,11 +1,0 @@
-<xml>
-<result type="int">1</result>
-<description type="text">Success</description>
-<error_no type="int">0</error_no>
-<error_message type="text">None</error_message>
-<key type="text">jul2009</key>
-<seq type="int">2</seq>
-<resource type="int">1</resource>
-<data type="text">http://scej-home.playstation.net/j.home/Home/sq/201009/scr_a_/tobira_HomesquareEvent-Sodium.dds</data>
-<status type="int">0</status>
-</xml>
